feat(DPI): enable optional file upload replacement in DPI FileUpload component
This commit adds a new option for enabling the replacement of uploaded files in the FileUpload component of the Data Provider Interface. If enabled, users can replace a file upload by clicking a 'replace' button that appears when editing the distribution. The code adds a new 'uploadOrReplaceFile' method that handles the upload or replacement, based on route parameters. The method finds the access URL index and passes it to the server when replacing a file. The commit also includes a helper method that extracts the file ID from the access URL of uploads. The change was also accompanied by minor UI updates.